# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ApiBileMo",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Api for Phones Catalog in B2B"
        }
    ],
    "paths": {
...
        "/phones": {
            "get": {
                "tags": [
                    "Phone"
                ],
                "operationId": "App\\Controller\\PhoneController::index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "/phones?page=2"
                    },
                    {
                        "name": "brand",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "/phones?brand=samsung"
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Syntaxe : /phones?price=[Xmin, Xmax] ou ?price=[Xmin]",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "/phones?price=[500,1000]"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste de téléphones mobiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PhonesList"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun téléphone pour cette recherche",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Aucun téléphone ne correspond aux crtières de recherche."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/phones/{id}": {
            "get": {
                "tags": [
                    "Phone"
                ],
                "operationId": "App\\Controller\\PhoneController::show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du téléphone",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Phone"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Security"
                ],
                "operationId": "App\\Controller\\SecurityController::register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inscription d'un utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "L'utilisateur a été enregistré avec succès !"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "User mal formaté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Le username et le mot de passe sont obligatoires."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login_check": {
            "post": {
                "tags": [
                    "Security"
                ],
                "operationId": "App\\Controller\\SecurityController::login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inscription d'un utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJpYXQiOjE2MDY4NTUyNzEsImV4cCI6MTYwNjg1ODg3MSwicm9sZXMiOlsiUk9MRV9VU0VSIl0sInVzZXJuYW1lIjoiQ2xpZW50LTEifQ.SVGKQhz-r9JLScuS3fzopa-oWKPDbv51ER9x_yILmvnLQcxX8DI2Yr75pR4iBR598GQtqkAYEsz1VO42CZICwXbgpibn9FrVzsxsyMm3AN4gJmlLaLVqoYL7aY1VSjPR5wMVoQQrKr9WqPBnGv4uo_CqzQiXByi8BBUYEtOg4pjcfNTQPnacjRwVa0eW4krgCJHKJfYX8v64oixk5pPUkHuHxJkdcQjqGzXQP5cnuS8jqhrOuEGVIKjqmLPfgXsPiMUXwfVigztzlF5-qpjAtKPFunkak15kh68Hu6Fif73UyXKK-aYqZ2_yMBlRsEpFav_AXtvFinpgZ2ZvmQXffWDe_f1Bm94ht_jYCz4zhr7n3UGRAsyyHAn7CmRopbG6ET4R-DVbGB4HHi9OMUeNTtSS02eUzPOhVNeVqzzh4mClapARvId1lbCcHuFjHuvRr-CMzE0nfkiHiRsXc-JHzHyAx9glnrOUre76-xUFH78NCm585gUv9HebCuI42n0neg_4XExDQ7gsRzaCNr-12JOrZowz66DyGw27QCrHyByRS1XKiVdrI2njwxbc0apkAAcsxBghzk9pDNcxYX69nfMK0TnePnD89kIZlBJz32i4TfSTdKB5tjc3ajiLGVKlukH1Oz1Q9H1gt6CXWcny7wpMXXgHRuumzkAMnnnxj0Q"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token périmé ou invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Le token est périmé ou invalide."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PhonesList": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "brand": {
                        "type": "string"
                    },
                    "model": {
                        "type": "string"
                    },
                    "price": {
                        "type": "integer"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "self": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Phone": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "brand": {
                        "type": "string"
                    },
                    "model": {
                        "type": "string"
                    },
                    "price": {
                        "type": "integer"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "self": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "NotFound": {
                "description": "La ressource n'existe pas.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "La ressource n'existe pas."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "ID de la ressource",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            }
        }
    }
}